requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.2,<3.0.0
aiohttp>=3.9.1,<4.0.0
asyncpg>=0.29.0,<1.0.0
aiofiles>=23.2.1,<24.0.0

# Database
alembic>=1.13.1,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0
sqlalchemy>=2.0.25,<3.0.0

# Security
pynacl>=1.5.0,<2.0.0

# Monitoring
prometheus-client>=0.19.0,<1.0.0

# Logging
python-json-logger>=2.0.7,<3.0.0

# Configuration
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.3,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

